+++ packages/generator-nick/generators/app/templates/knexfile.js
@@ -1,0 +1,26 @@
+var path = require('path');
+var { config } = require('./config');
+
+const knexSettings = {
+  client: 'pg',
+  connection: config.connection,
+  pool: {
+    min: 2,
+    max: 10,
+  },
+  migrations: {
+    directory: [
+      path.resolve(__dirname, './src/develop/nick/src/migrations'),
+      path.resolve(__dirname, './src/migrations'),
+    ],
+    tableName: 'knex_migrations',
+  },
+  seeds: {
+    directory: path.resolve(__dirname, './src/develop/nick/src/seeds'),
+  },
+};
+
+module.exports = {
+  development: knexSettings,
+  production: knexSettings,
+};